# api.yml
#Test with: http://editor.swagger.io/

swagger: '2.0'
info:
  version: 1.0.0
  title: Microservice for Events
  description: Go + Postgres
  license:
    name: Eventum Team
host: host:8000
basePath: /ev-events-ms/app
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /events:
    post:
      summary: Creates a new event
      operationId: createEvent
      parameters:
        - name: event
          in: body
          required: true
          description: Event Object
          schema:
            $ref: '#/definitions/EventInput'
      responses:
        '201':
          description: Event Created Successfully
          schema:
            $ref: '#/definitions/MessageModel'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    get:
      summary: Returns all events
      operationId: getAllEvents
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EventsList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /events/{id}:
    get:
      summary: Returns an event with the given id
      operationId: getEventById
      parameters:
        - name: id
          in: path
          description: The id of the event to be returned
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Event'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'

    put:
      summary: Updates an event with the given id
      operationId: updateEvent
      parameters:
        - name: id
          in: path
          description: The id of the event to be updated
          required: true
          type: integer
        - name: event
          in: body
          required: true
          description: Event Object
          schema:
            $ref: '#/definitions/EventInput'
      responses:
        '200':
          description: ok
          schema:
            $ref: '#/definitions/MessageModel'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'

    delete:
      summary: Deletes an event with the given id
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          description: The id of the event to be deleted
          required: true
          type: integer
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'


  /events/filter/locations:
    get:
      summary: Returns all events locations
      operationId: getLocations
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/LocationsList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /events/filter/status:
    get:
      summary: Returns all events given a status. ?status=(active, cancelled, ended)
      operationId: getEventsByStatus
      parameters:
        - name: status
          in: path
          description: (active, cancelled, ended)
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EventsList'
        '400':
          description: Status Bad Request
          schema:
            $ref: '#/definitions/EventsList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /events/filter/ownerType:
    get:
      summary: Returns all events given a owner type. ?type=(group, user).
      operationId: getEventsByOwnerType
      parameters:
        - name: type
          in: path
          description: (group, user)
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EventsList'
        '400':
          description: Status Bad Request
          schema:
            $ref: '#/definitions/EventsList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /events/filter/name:
    get:
      summary: Returns all events where the string submited is part of their names. ?name=(string)
      operationId: getEventsByName
      parameters:
        - name: name
          in: path
          description: provide a substring of an event name
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EventsList'
        '400':
          description: Status Bad Request
          schema:
            $ref: '#/definitions/EventsList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /events/filter/rangeDate:
    get:
      summary: Returns all events given range date. ?start&end
      operationId: getEventsByRangeDate
      parameters:
        - name: name
          in: path
          description: date format "2000-01-01T00:00:00Z"
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EventsList'
        '400':
          description: Status Bad Request
          schema:
            $ref: '#/definitions/EventsList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

definitions:
  # Events
  Event:
    type: object
    properties:
      id:
        type: string
      ownerId:
        type: string
      ownerType:
        type: string
      name:
        type: string
      status:
        type: string
      eventStartDate:
        type: string
      eventFinishDate:
        type: string
      description:
        type: string
      url:
        type: string
      latitude:
        type: string
      longitude:
        type: string

  EventInput:
    type: object
    properties:
      ownerId:
        type: string
      ownerType:
        type: string
      name:
        type: string
      status:
        type: string
      eventStartDate:
        type: string
      eventFinishDate:
        type: string
      description:
        type: string
      url:
        type: string
      latitude:
        type: string
      longitude:
        type: string

  EventsList:
    type: object
    required:
      - total
      - list
    properties:
      total:
        type: number
      list:
        type: array
        items:
          $ref: '#/definitions/Event'

  Location:
    type: object
    properties:
      eventId:
        type: string
      eventName:
        type: string
      latitude:
        type: string
      longitude:
        type: string

  LocationsList:
    type: object
    required:
      - total
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/Location'

  # Errors
  ErrorModel:
    type: object
    required:
      - message
      - code
    properties:
      error:
        type: string
      code:
        type: number
      description:
        type: object

  MessageModel:
    type: object
    required:
      - message
      - code
    properties:
      error:
        type: string
      code:
        type: number
      description:
        type: object